refactor: extract attribution calculation policy
